add temp vs screen, add titles folder, add instructions
