+++ BeanWare/Assets/Scenes/Titles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a2f1a341882e2c948a3d185a638ccb5f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
