+++ BeanWare/Assets/Scenes/Titles/InstructionsScreen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a91d01bb734340d47853be515a7764b1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
