+++ BeanWare/Assets/Scenes/Titles/TitleScreen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0bbce79cf07ae7849b680330b199d4da
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
