+++ BeanWare/Assets/Scenes/Titles/VersusScreen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e29455e603f3c67478ceb494af030006
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
