+++ BeanWare/Assets/Scripts/ScreenLoaders.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: bdedd1fc276d9264a8d670dd4bb9de2f
+guid: 29c26bbb9153a3b478e1ca8bde9ccb64
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
