+++ BeanWare/Assets/Scripts/ScreenLoaders/InstructionsLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aef7d4f96892a02408e5dcaa43c57028
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
