+++ BeanWare/Assets/Scripts/ScreenLoaders/TextAnimator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df124a000768dfb4887b0bafb7319df7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
