+++ BeanWare/Assets/Scripts/ScreenLoaders/VersusScreenLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d69490d787623e14e9ef2318fc461e98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
